needs to be fixed

+++ includes/structs/ray.h
@@ -6,7 +6,7 @@
 /*   By: konsolka <konsolka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/27 21:38:59 by user              #+#    #+#             */
-/*   Updated: 2020/04/28 18:47:53 by konsolka         ###   ########.fr       */
+/*   Updated: 2020/04/28 19:42:58 by konsolka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct	s_ray
 	double		len;
 	float		pov;
 	int			wall_c;
-	int			angle;
+	double		angle;
 }				t_ray;
 
 #endif
